Use size_t for sizes

+++ hw1/memPage_t.h
@@ -5,7 +5,7 @@ class memPage_t {
 public:
 	
 	// Constructs with a given maximum capacity
-	memPage_t(unsigned int capacity) :
+	memPage_t(size_t capacity) :
 		capacity(capacity), size(0), store(new char[capacity]), pos(0) {}
 	~memPage_t() { delete[] store; }
 
@@ -20,21 +20,21 @@ public:
 	void set_pos(unsigned int new_pos) { pos = new_pos; }  // TODO handle OOB index
 
 	// Access
-	void read(void* out, size_t len, unsigned int pos = pos) const;
-	void write(void* const in, size_t len, unsigned int pos = pos);
+	void read(void* out, size_t len, size_t pos = pos) const;
+	void write(void* const in, size_t len, size_t pos = pos);
 
 private:
 	// Disallow copy constructor
 	memPage_t(const memPage_t& that);
 
 	// Max capacity
-	unsigned int capacity;
+	size_t capacity;
 	// Actual size
-	unsigned int size;
+	size_t size;
 	// Backing store
 	char* store;
 	// Cursor
-	unsigned int pos;
+	size_t pos;
 };
 
 #endif
